udevtrigger: trigger devices w/o dev sysfs entry
A lot of devices aren't represented via major/minor numbers
and thus don't have dev entry in their sysfs directory, like
network devices, USB, power supplies etc.
It looks to be the rigth thing to trigger them as well.

Anyway, it's up to the hotplug daemon to decide what to do with devices,
it needs full vision of what's present in the system.

Signed-off-by: Sergiy Kibrik <sakib@meta.ua>